+++ .vs/VSWorkspaceState.json
@@ -2,6 +2,6 @@
   "ExpandedNodes": [
     ""
   ],
-  "SelectedNode": "\\H:\\Students_Folder\\Charlie Hatch\\Semesters\\Multi-Semester\\App Dev\\Projects\\App Dev 1\\TicketingSystem",
+  "SelectedNode": "\\TicketingSystem.sln",
   "PreviewInSolutionExplorer": false
 }
