+++ Program.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace TicketingSystem
+namespace TicketingApp
 {
     internal static class Program
     {
