+++ TicketingApp.csproj
@@ -4,10 +4,10 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{4077BF1D-E0BB-4359-9C0C-529D756C64DE}</ProjectGuid>
+    <ProjectGuid>{F9257137-7BF6-47DF-A356-0AF1CA1BAFAD}</ProjectGuid>
     <OutputType>WinExe</OutputType>
-    <RootNamespace>TicketingSystem</RootNamespace>
-    <AssemblyName>TicketingSystem</AssemblyName>
+    <RootNamespace>TicketingApp</RootNamespace>
+    <AssemblyName>TicketingApp</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -52,8 +52,12 @@
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
+    <Compile Include="NewTicket.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <EmbeddedResource Include="Form1.resx">
+      <DependentUpon>Form1.cs</DependentUpon>
+    </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
