+++ TicketingApp.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.8.34511.84
+VisualStudioVersion = 17.9.34607.119
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TicketingApp", "TicketingApp.csproj", "{4077BF1D-E0BB-4359-9C0C-529D756C64DE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TicketingApp", "TicketingApp.csproj", "{F9257137-7BF6-47DF-A356-0AF1CA1BAFAD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4077BF1D-E0BB-4359-9C0C-529D756C64DE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4077BF1D-E0BB-4359-9C0C-529D756C64DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4077BF1D-E0BB-4359-9C0C-529D756C64DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4077BF1D-E0BB-4359-9C0C-529D756C64DE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F9257137-7BF6-47DF-A356-0AF1CA1BAFAD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F9257137-7BF6-47DF-A356-0AF1CA1BAFAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F9257137-7BF6-47DF-A356-0AF1CA1BAFAD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F9257137-7BF6-47DF-A356-0AF1CA1BAFAD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E3E8CE21-24ED-47EC-ADFD-0AF65E7E6836}
+		SolutionGuid = {03C395F1-9DE4-469C-9214-D67E07871EB3}
 	EndGlobalSection
 EndGlobal
